feat: Finish totp implementation
