+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
 
+- Add TOTP recipe
+
 ## [4.4.1] - 2023-03-09
 
 - Normalises email in all APIs in which email was not being
