+++ src/main/java/io/supertokens/webserver/api/totp/VerifyTotpDeviceAPI.java
@@ -68,7 +68,7 @@ public class VerifyTotpDeviceAPI extends WebserverAPI {
             super.sendJsonResponse(200, result, resp);
         } catch (LimitReachedException e) {
             result.addProperty("status", "LIMIT_REACHED_ERROR");
-            result.addProperty("retryAfterSec", e.retryInSeconds);
+            result.addProperty("retryAfterMs", e.retryInMs);
             super.sendJsonResponse(200, result, resp);
         } catch (StorageQueryException | StorageTransactionLogicException e) {
             throw new ServletException(e);
