+++ src/test/java/io/supertokens/test/totp/api/VerifyTotpAPITest.java
@@ -1,9 +1,12 @@
-package io.supertokens.test.totp.totp;
+package io.supertokens.test.totp.api;
 
 import com.google.gson.JsonObject;
+
 import io.supertokens.ProcessState;
 import io.supertokens.test.httpRequest.HttpResponseException;
+import io.supertokens.test.totp.TOTPRecipeTest;
 import io.supertokens.pluginInterface.STORAGE_TYPE;
+import io.supertokens.pluginInterface.totp.TOTPDevice;
 import io.supertokens.storageLayer.StorageLayer;
 import io.supertokens.test.TestingProcessManager;
 import io.supertokens.test.Utils;
@@ -16,7 +19,7 @@ import org.junit.rules.TestRule;
 
 import static org.junit.Assert.*;
 
-public class RemoveTotpDeviceAPITest {
+public class VerifyTotpAPITest {
 
     @Rule
     public TestRule watchman = Utils.getOnFailure();
@@ -31,13 +34,13 @@ public class RemoveTotpDeviceAPITest {
         Utils.reset();
     }
 
-    private Exception removeDeviceRequest(TestingProcessManager.TestingProcess process, JsonObject body) {
+    private Exception updateDeviceRequest(TestingProcessManager.TestingProcess process, JsonObject body) {
         return assertThrows(
                 io.supertokens.test.httpRequest.HttpResponseException.class,
                 () -> HttpRequestForTesting.sendJsonPOSTRequest(
                         process.getProcess(),
                         "",
-                        "http://localhost:3567/recipe/totp/device/remove",
+                        "http://localhost:3567/recipe/totp/verify",
                         body,
                         1000,
                         1000,
@@ -65,6 +68,11 @@ public class RemoveTotpDeviceAPITest {
     public void testApi() throws Exception {
         String[] args = { "../" };
 
+        // Trigger rate limiting on 1 wrong attempts:
+        Utils.setValueInConfig("totp_max_attempts", "1");
+        // Set cooldown to 1 second:
+        Utils.setValueInConfig("totp_rate_limit_cooldown_sec", "1");
+
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
 
@@ -75,7 +83,7 @@ public class RemoveTotpDeviceAPITest {
         // Setup user and devices:
         JsonObject createDeviceReq = new JsonObject();
         createDeviceReq.addProperty("userId", "user-id");
-        createDeviceReq.addProperty("deviceName", "d1");
+        createDeviceReq.addProperty("deviceName", "deviceName");
         createDeviceReq.addProperty("period", 30);
         createDeviceReq.addProperty("skew", 0);
 
@@ -90,53 +98,86 @@ public class RemoveTotpDeviceAPITest {
                 Utils.getCdiVersionLatestForTests(),
                 "totp");
         assertEquals(createDeviceRes.get("status").getAsString(), "OK");
+        String secretKey = createDeviceRes.get("secret").getAsString();
 
-        // create another device d2:
-        createDeviceReq.addProperty("deviceName", "d2");
-        JsonObject createDeviceRes2 = HttpRequestForTesting.sendJsonPOSTRequest(
-                process.getProcess(),
-                "",
-                "http://localhost:3567/recipe/totp/device",
-                createDeviceReq,
-                1000,
-                1000,
-                null,
-                Utils.getCdiVersionLatestForTests(),
-                "totp");
-        assertEquals(createDeviceRes2.get("status").getAsString(), "OK");
+        TOTPDevice device = new TOTPDevice("user-id", "deviceName", secretKey, 30, 0, false);
 
-        // Start the actual tests for remove device API:
+        // Start the actual tests for update device API:
 
         JsonObject body = new JsonObject();
 
-        // Missing userId/deviceName
+        // Missing userId/deviceName/skew/period
         {
-            Exception e = removeDeviceRequest(process, body);
+            Exception e = updateDeviceRequest(process, body);
             checkFieldMissingErrorResponse(e, "userId");
 
             body.addProperty("userId", "");
-            e = removeDeviceRequest(process, body);
-            checkFieldMissingErrorResponse(e, "deviceName");
+            e = updateDeviceRequest(process, body);
+            checkFieldMissingErrorResponse(e, "totp");
 
+            body.addProperty("totp", "");
+            e = updateDeviceRequest(process, body);
+            checkFieldMissingErrorResponse(e, "allowUnverifiedDevices");
         }
 
-        // Invalid userId/deviceName
+        // Invalid userId/deviceName/skew/period
         {
-            body.addProperty("deviceName", "");
-            Exception e = removeDeviceRequest(process, body);
+            body.addProperty("allowUnverifiedDevices", true);
+            Exception e = updateDeviceRequest(process, body);
             checkResponseErrorContains(e, "userId cannot be empty"); // Note that this is not a field missing error
 
-            body.addProperty("userId", "user-id");
-            e = removeDeviceRequest(process, body);
-            checkResponseErrorContains(e, "deviceName cannot be empty");
+            body.addProperty("userId", device.userId);
+            e = updateDeviceRequest(process, body);
+            checkResponseErrorContains(e, "totp must be 6 characters long");
 
-            body.addProperty("deviceName", "d1");
+            // test totp of length 5:
+            body.addProperty("totp", "12345");
+            e = updateDeviceRequest(process, body);
+            checkResponseErrorContains(e, "totp must be 6 characters long");
 
-            // should pass now:
+            // test totp of length 8:
+            body.addProperty("totp", "12345678");
+            e = updateDeviceRequest(process, body);
+            checkResponseErrorContains(e, "totp must be 6 characters long");
+
+            // but let's pass invalid code first
+            body.addProperty("totp", "123456");
+            JsonObject res0 = HttpRequestForTesting.sendJsonPOSTRequest(
+                    process.getProcess(),
+                    "",
+                    "http://localhost:3567/recipe/totp/verify",
+                    body,
+                    1000,
+                    1000,
+                    null,
+                    Utils.getCdiVersionLatestForTests(),
+                    "totp");
+            assert res0.get("status").getAsString().equals("INVALID_TOTP_ERROR");
+
+            // Check that rate limiting is triggered for the user:
+            JsonObject res3 = HttpRequestForTesting.sendJsonPOSTRequest(
+                    process.getProcess(),
+                    "",
+                    "http://localhost:3567/recipe/totp/verify",
+                    body,
+                    1000,
+                    1000,
+                    null,
+                    Utils.getCdiVersionLatestForTests(),
+                    "totp");
+            assert res3.get("status").getAsString().equals("LIMIT_REACHED_ERROR");
+            assert res3.get("retryAfterSec") != null;
+
+            // wait for cooldown to end (1s)
+            Thread.sleep(1000);
+
+            // should pass now on valid code
+            String validTotp = TOTPRecipeTest.generateTotpCode(process.getProcess(), device);
+            body.addProperty("totp", validTotp);
             JsonObject res = HttpRequestForTesting.sendJsonPOSTRequest(
                     process.getProcess(),
                     "",
-                    "http://localhost:3567/recipe/totp/device/remove",
+                    "http://localhost:3567/recipe/totp/verify",
                     body,
                     1000,
                     1000,
@@ -144,35 +185,34 @@ public class RemoveTotpDeviceAPITest {
                     Utils.getCdiVersionLatestForTests(),
                     "totp");
             assert res.get("status").getAsString().equals("OK");
-            assert res.get("didDeviceExist").getAsBoolean() == true;
 
-            // try again with same device (still pass but didDeviceExist should be false)
+            // try to reuse the same code (replay attack)
+            body.addProperty("totp", "mycode");
             JsonObject res2 = HttpRequestForTesting.sendJsonPOSTRequest(
                     process.getProcess(),
                     "",
-                    "http://localhost:3567/recipe/totp/device/remove",
+                    "http://localhost:3567/recipe/totp/verify",
                     body,
                     1000,
                     1000,
                     null,
                     Utils.getCdiVersionLatestForTests(),
                     "totp");
-            assert res2.get("status").getAsString().equals("OK");
-            assert res2.get("didDeviceExist").getAsBoolean() == false;
+            assert res2.get("status").getAsString().equals("INVALID_TOTP_ERROR");
 
-            // try deleting device for a non-existent user
+            // try verifying device for a non-existent user
             body.addProperty("userId", "non-existent-user");
-            JsonObject res3 = HttpRequestForTesting.sendJsonPOSTRequest(
+            JsonObject res5 = HttpRequestForTesting.sendJsonPOSTRequest(
                     process.getProcess(),
                     "",
-                    "http://localhost:3567/recipe/totp/device/remove",
+                    "http://localhost:3567/recipe/totp/verify",
                     body,
                     1000,
                     1000,
                     null,
                     Utils.getCdiVersionLatestForTests(),
                     "totp");
-            assert res3.get("status").getAsString().equals("TOTP_NOT_ENABLED_ERROR");
+            assert res5.get("status").getAsString().equals("TOTP_NOT_ENABLED_ERROR");
         }
 
         process.kill();
